pyproject.toml
[project]
name = "livekitx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.0.25",
    "livekit-api>=1.0.0",
    "livekit-agents>=1.3.11",
    "livekit-plugins-deepgram>=1.3.11",
    "livekit-plugins-elevenlabs>=1.3.11",
    "livekit-plugins-openai>=1.3.11",
    "livekit-plugins-silero>=1.3.11",
    "python-dotenv>=1.2.1",
]
